fixed merge conflicts

+++ src/main/java/com/techelevator/model/dao/BeerReviewDAO.java
@@ -9,6 +9,6 @@ public interface BeerReviewDAO {
 
     public List<BeerReview> getReviewsByBeerId (int beerId);
     public BeerReview getReviewById (int id);
-    public void createNewReview(int beerId, int userId, double rating, String review,
+    public void createNewReview(int beerId, int userId, int rating, String review,
                                 String reviewTitle);
 }
